@@ -1,15 +1,8 @@
 import { Component } from "@angular/core";
-import { SwPush, SwUpdate } from "@angular/service-worker";
-import { interval } from "rxjs";
-import { HttpClient } from "@angular/common/http";
 
 @Component({
   selector: "app-root",
   templateUrl: "./app.component.html",
   styleUrls: ["./app.component.css"],
 })
-export class AppComponent {
-  constructor() {
-    console.log("test");
-  }
-}
+export class AppComponent {}
