@@ -1,14 +1,19 @@
-import { ModuleWithProviders }  from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import {ForecastsListComponent} from "./forecasts-list/forecasts-list.component";
-import {MainPageComponent} from "./main-page/main-page.component";
+import { ModuleWithProviders } from "@angular/core";
+import { Routes, RouterModule } from "@angular/router";
+import { ForecastsListComponent } from "./forecasts-list/forecasts-list.component";
+import { MainPageComponent } from "./main-page/main-page.component";
 
 const appRoutes: Routes = [
   {
-    path: '', component: MainPageComponent
+    path: "",
+    component: MainPageComponent,
   },
   {
-    path: 'forecast/:zipcode', component: ForecastsListComponent
-  }
+    path: "forecast/:zipcode",
+    component: ForecastsListComponent,
+  },
 ];
-export const routing: ModuleWithProviders<any> = RouterModule.forRoot(appRoutes, {});
+export const routing: ModuleWithProviders<any> = RouterModule.forRoot(
+  appRoutes,
+  { bindToComponentInputs: true }
+);
